Add quantized reshape version without oneDNN support.
