+++ src/operator/quantization/dnnl/dnnl_quantized_reshape.cc
@@ -23,86 +23,48 @@
  */
 
 #if MXNET_USE_ONEDNN == 1
-#include "./dnnl_quantized_reshape-inl.h"
+#include "operator/quantization/quantized_reshape-inl.h"
+#include "operator/nn/dnnl/dnnl_ops-inl.h"
 
 namespace mxnet {
 namespace op {
 
-DMLC_REGISTER_PARAMETER(QuantizedReshapeParam);
+static void DNNLQuantizedReshapeForward(const nnvm::NodeAttrs& attrs,
+                                        const OpContext& ctx,
+                                        const std::vector<NDArray>& inputs,
+                                        const std::vector<OpReqType>& req,
+                                        const std::vector<NDArray>& outputs) {
+  CHECK(inputs[0].dtype() == mshadow::kUint8 || inputs[0].dtype() == mshadow::kInt8)
+      << "dnnl_quantized_reshape op only supports uint8 and int8 as input type";
 
-NNVM_REGISTER_OP(_contrib_quantized_reshape)
-    .add_alias("_npx_quantized_reshape")
-    .set_num_inputs(3)
-    .set_num_outputs(3)
-    .set_attr_parser(ParamParser<QuantizedReshapeParam>)
-    .set_attr<nnvm::FListInputNames>(
-        "FListInputNames",
-        [](const NodeAttrs& attrs) {
-          return std::vector<std::string>{"data", "min_data", "max_data"};
-        })
-    .set_attr<nnvm::FListOutputNames>(
-        "FListOutputNames",
-        [](const NodeAttrs& attrs) {
-          return std::vector<std::string>{"output", "min_output", "max_output"};
-        })
-    .set_attr<nnvm::FInplaceOption>(
-        "FInplaceOption",
-        [](const NodeAttrs& attrs) {
-          return std::vector<std::pair<int, int> >{{0, 0}, {1, 1}, {2, 2}};
-        })
-    .set_attr<FComputeEx>("FComputeEx<cpu>", DNNLQuantizedReshapeForward)
-    .set_attr<FResourceRequest>("FResourceRequest",
-                                [](const NodeAttrs& n) {
-                                  return std::vector<ResourceRequest>{ResourceRequest::kTempSpace};
-                                })
-    .set_attr<FInferStorageType>("FInferStorageType", QuantizedReshapeStorageType)
-    .set_attr<mxnet::FInferShape>("FInferShape", QuantizedReshapeInferShape)
-    .set_attr<nnvm::FInferType>("FInferType", QuantizedReshapeType)
-    .set_attr<nnvm::FGradient>("FGradient", MakeZeroGradNodes)
-    .set_attr<FQuantizable>("FQuantizable",
-                            [](const NodeAttrs& attrs) { return QuantizeType::kSupport; })
-    .add_argument("data", "NDArray-or-Symbol", "Array to be reshaped.")
-    .add_argument("min_data",
-                  "NDArray-or-Symbol",
-                  "The minimum scalar value "
-                  "possibly produced for the data")
-    .add_argument("max_data",
-                  "NDArray-or-Symbol",
-                  "The maximum scalar value "
-                  "possibly produced for the data")
-    .add_arguments(QuantizedReshapeParam::__FIELDS__());
-
-template <bool is_numpy_op>
-nnvm::ObjectPtr QuantizedReshapeNode(const NodeAttrs& attrs) {
-  QuantizedReshapeParam param;
-  if (is_numpy_op) {
-    const NumpyXReshapeParam& _param = nnvm::get<NumpyXReshapeParam>(attrs.parsed);
-    param.newshape                   = _param.newshape;
-    param.reverse                    = _param.reverse;
-    param.order                      = _param.order;
-    param.keep_highest               = false;
-    param.is_numpy_op                = true;
+  if (SupportDNNLReshape(inputs[0], outputs[0])) {
+    OpReqType reqType;
+    if (inputs[0].GetDNNLData()->get_data_handle() != outputs[0].GetDNNLData()->get_data_handle())
+      reqType = kWriteTo;
+    else
+      reqType = req[0];
+    DNNLRun(DNNLReshapeForward, attrs, ctx, inputs[0], reqType, outputs[0]);
   } else {
-    const ReshapeParam& _param = nnvm::get<ReshapeParam>(attrs.parsed);
-    param.shape                = _param.shape;
-    param.keep_highest         = _param.keep_highest;
-    param.reverse              = _param.reverse;
-    param.is_numpy_op          = false;
+    FallBackCompute(UnaryOp::IdentityCompute<cpu>, attrs, ctx, inputs, req, outputs);
   }
 
-  nnvm::ObjectPtr node = nnvm::Node::Create();
-  node->attrs.op       = Op::Get("_contrib_quantized_reshape");
-  node->attrs.name     = "quantized_" + attrs.name;
-  param.SetAttrDict(&(node->attrs.dict));
-  if (node->op() != nullptr && node->op()->attr_parser != nullptr) {
-    node->op()->attr_parser(&(node->attrs));
-  }
-  return node;
+  *outputs[1].data().dptr<float>() = *inputs[1].data().dptr<float>();
+  *outputs[2].data().dptr<float>() = *inputs[2].data().dptr<float>();
 }
 
-NNVM_REGISTER_OP(_npx_reshape).set_attr<FQuantizedOp>("FQuantizedOp", QuantizedReshapeNode<true>);
+inline bool QuantizedReshapeStorageType(const nnvm::NodeAttrs& attrs,
+                                        const int dev_mask,
+                                        DispatchMode* dispatch_mode,
+                                        std::vector<int>* in_attrs,
+                                        std::vector<int>* out_attrs) {
+  CHECK_EQ(in_attrs->size(), 3U);
+  CHECK_EQ(out_attrs->size(), 3U);
+  return DNNLStorageType(attrs, dev_mask, true, dispatch_mode, in_attrs, out_attrs);
+}
 
-NNVM_REGISTER_OP(Reshape).set_attr<FQuantizedOp>("FQuantizedOp", QuantizedReshapeNode<false>);
+NNVM_REGISTER_OP(_contrib_quantized_reshape)
+    .set_attr<FComputeEx>("FComputeEx<cpu>", DNNLQuantizedReshapeForward)
+    .set_attr<FInferStorageType>("FInferStorageType", QuantizedReshapeStorageType);
 
 }  // namespace op
 }  // namespace mxnet
